requirements.txt
Authlib==0.15.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.4.3
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
pycparser==2.20
python-dotenv==0.15.0
requests==2.25.1
urllib3==1.26.3
Werkzeug==1.0.1
